added userController et modif aux ficher UserServiceImpl et UserServiceInter

+++ src/main/java/fr/erickfranco/cv_api/services/serviceimpl/UserServiceImpl.java
@@ -1,16 +1,19 @@
 package fr.erickfranco.cv_api.services.serviceimpl;
 
+import fr.erickfranco.cv_api.configurations.exceptionconfig.exception.NotFoundExcepton;
 import fr.erickfranco.cv_api.models.User;
 import fr.erickfranco.cv_api.repositories.UserRepository;
 import fr.erickfranco.cv_api.services.serviceinter.UserServiceInter;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
 /**
  * @author Erick Franco
@@ -41,17 +44,29 @@ public class UserServiceImpl implements UserServiceInter {
     }
 
     @Override
-    public Optional<User> findUserById(Long id) {
-        return Optional.empty();
+    public User findUserById(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new NotFoundExcepton("l'utilisateur avec l'id " + id + " n'existe pas ");
+        }
+        return userRepository.getOne(id);
     }
 
     @Override
     public User saveUser(User user) {
+        user.setPassword(passwordEncoder().encode(user.getPassword()));
         return userRepository.save(user);
+    }
+
+    @Bean
+    private PasswordEncoder passwordEncoder(){
+        return  new BCryptPasswordEncoder();
     }
 
     @Override
     public void deleteById(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new NotFoundExcepton("l'utilisateur que vous souhaitez l'eliminer avec l'id numéro " + id + " n'existe pas ");
+        }
         userRepository.deleteById(id);
     }
 }
